Refactored and simplified AudioBuffer classes.
