+++ cpvc/cpvc.csproj
@@ -138,8 +138,6 @@
     <Compile Include="Machine\Interfaces\IReversibleMachine.cs" />
     <Compile Include="Machine\MachineHistory.cs" />
     <Compile Include="Miscellaneous\AudioBuffer.cs" />
-    <Compile Include="Miscellaneous\BaseAudioBuffer.cs" />
-    <Compile Include="Miscellaneous\CircularAudioBuffer.cs" />
     <Compile Include="Miscellaneous\IBlob.cs" />
     <Compile Include="Machine\Interfaces\IBookmarkableMachine.cs" />
     <Compile Include="Machine\Interfaces\IInteractiveMachine.cs" />
